third_party/WebKit/LayoutTests/crypto/encrypt-decrypt-expected.txt

Tests cypto.subtle.encrypt and crypto.subtle.decrypt

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS crypto.subtle.encrypt({name: 'AES-CBC', iv: null}, keys.aesCbc, data) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS crypto.subtle.encrypt({name: 'AES-CBC'}, keys.aesCbc, data) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS crypto.subtle.encrypt({name: 'AES-CBC', iv: 3}, keys.aesCbc, data) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS crypto.subtle.encrypt({name: 'AES-CBC', iv: new Uint8Array[0]}, keys.aesCbc, data) threw exception TypeError: undefined is not a function.
PASS crypto.subtle.encrypt(aesCbc, keys.hmacSha1, data) threw exception NotSupportedError: key.algorithm does not match that of operation.
PASS crypto.subtle.encrypt(aesCbc, keys.aesCbcJustDecrypt, data) threw exception NotSupportedError: key.usages does not permit this operation.
PASS successfullyParsed is true

TEST COMPLETE
